Calcula número de bancas legislativas a repartir en una elección

Función que calcula la distribución proporcional de escaños para la categoría Diputado Nacional y de mayoría/minoría para Senaor Nacional en función de votos obtenidos

(Function that computes propotional allocation of Diputados and Senadores seats)

compute_seats(data)

Arguments

data

un data.frame con los resultados de una elección para agregados a nivel provincial

(a data.frame with aggregate electoral results at provincial level).

Details

(Computes allocation of legislative seats)

La distribución de escaños esta regida por la formula del sistema D'Hondt para Diputados y mayoría/minoría para Senadores.

La cantidad de escaños de cada provincia dependen de su población con un mínimo de \(5\) por provincia. En caso de Senadores se asignan \(2\) al de mayor votos y \(1\) al segundo.

En el caso de Diputados, La renovación de bancas de cada provincia serealiza por mitades cada dos años. Cuando la cantidad de escaños que corresponden a una provincia es impar las mismas eligen un diputado más en uno de los turnos: o concurrentes con elecciones presidenciales, o en elecciones de mitad de termino presidencial. En el caso de Senadores su mandato es de 6 años y se renuevan parcialmente por grupos de 8 provincias.

See also

Examples

electorAr::get_election_data(district = "caba",
                          category = "dip",
                          round = "gral",
                          year = 2007) -> caba_dip_2007

 caba_dip_2007
#> # A tibble: 32 × 8
#> # Groups:   codprov [1]
#>    category round  year codprov name_prov electores listas  votos
#>    <chr>    <chr> <dbl> <chr>   <chr>         <dbl> <chr>   <dbl>
#>  1 dip      gral   2007 01      CABA        2564950 0003   104599
#>  2 dip      gral   2007 01      CABA        2564950 0005     4634
#>  3 dip      gral   2007 01      CABA        2564950 0009    79154
#>  4 dip      gral   2007 01      CABA        2564950 0014    12107
#>  5 dip      gral   2007 01      CABA        2564950 0020    41071
#>  6 dip      gral   2007 01      CABA        2564950 0023     4409
#>  7 dip      gral   2007 01      CABA        2564950 0036    26889
#>  8 dip      gral   2007 01      CABA        2564950 0037     3177
#>  9 dip      gral   2007 01      CABA        2564950 0038    33654
#> 10 dip      gral   2007 01      CABA        2564950 0050   247836
#> # ℹ 22 more rows

 compute_seats(data = caba_dip_2007)
#> # A tibble: 7 × 8
#> # Groups:   year, category, round, codprov, name_prov, listas [7]
#>   codprov name_prov  year category round listas nombre_lista               seats
#>   <chr>   <chr>     <dbl> <chr>    <chr> <chr>  <chr>                      <int>
#> 1 01      CABA       2007 dip      gral  0252   Al. Coalición Cívica           3
#> 2 01      CABA       2007 dip      gral  0050   Partido Socialista             2
#> 3 01      CABA       2007 dip      gral  0134   Alianza Frente para la Vi…     2
#> 4 01      CABA       2007 dip      gral  0136   Alianza Propuesta Republi…     2
#> 5 01      CABA       2007 dip      gral  0003   Unión Cívica Radical           1
#> 6 01      CABA       2007 dip      gral  0137   Alianza Diálogo por Bueno…     1
#> 7 01      CABA       2007 dip      gral  0138   Alianza Proyecto Sur           1