Descarga resultados electorales (Download electoral results)

Función que descarga resultados electorales agregados, nacionales y provinciales desde 2007

(Function that downloads national electoral data since 2007)

get_election_results(
  district = NULL,
  category = NULL,
  round = NULL,
  year = NULL
)

Arguments

district

un character con código para Argentina y las 24 provincias

(a named character with code for Argentina and the 24 provinces).

category

un character para la categoría electoral: diputado dip, senador sen o presidente presi

(a character with a name for the electoral category: deputy dip, senator sen or president presi).

round

un character para tipo de elección: primaria paso o general gral

(a character with a name for the election round: primary paso or general gral).

year

un integer para el año de eleccion

(an integer for the year of choice).

Value

devuelve un tibble con class "spec_tbl_df", "tbl_df","tbl", "data.frame" con los resultados de una elección determinada

(returns a tibble with electoral results of interest with class "spec_tbl_df", "tbl_df","tbl", "data.frame").

Examples

 get_election_data(district = "caba",
                   category = "dip",
                   round = "paso",
                   year = 2011,
                   level = "provincia",
                   long = TRUE, raw = FALSE)
#> # A tibble: 14 × 8
#> # Groups:   codprov [1]
#>    category round  year codprov name_prov electores listas   votos
#>    <chr>    <chr> <dbl> <chr>   <chr>         <dbl> <chr>    <dbl>
#>  1 dip      paso   2011 01      CABA        2491166 0023    186276
#>  2 dip      paso   2011 01      CABA        2491166 0036     68575
#>  3 dip      paso   2011 01      CABA        2491166 0179      8234
#>  4 dip      paso   2011 01      CABA        2491166 0302     68001
#>  5 dip      paso   2011 01      CABA        2491166 0501    509590
#>  6 dip      paso   2011 01      CABA        2491166 0504     95714
#>  7 dip      paso   2011 01      CABA        2491166 0508     47956
#>  8 dip      paso   2011 01      CABA        2491166 0509    145580
#>  9 dip      paso   2011 01      CABA        2491166 0510    201173
#> 10 dip      paso   2011 01      CABA        2491166 0517     81666
#> 11 dip      paso   2011 01      CABA        2491166 0518    294312
#> 12 dip      paso   2011 01      CABA        2491166 0536    135361
#> 13 dip      paso   2011 01      CABA        2491166 blancos  35478
#> 14 dip      paso   2011 01      CABA        2491166 nulos    27459