Obtiene nombres de listas (Get party names)

Función que agrega el nombre de las listas o partidos como columna a un tibble obtenido con get_election_data(long = TRUE) (Function that adds party labels as a column to a tibble obtained with get_election_data(long = TRUE)).

get_names(data)

Arguments

data

un tibble descargado con get_election_data(long = TRUE) guardado como objeto en el Enviroment (A get_election_data(long = TRUE) tibble saved as an object in the Enviroment).

Value

Devuelve el data set original con una columna extra con la identifiacion de las listas o partidos politicos. Como el objeto de entrada, este es class "tbl_df","tbl","data.frame"

(it retruns the original data set with a binded column with political parties names. As the original input the object is of class "tbl_df","tbl","data.frame").

Details

El formato de data debe ser long para poder obtener nombres de listas con get_names. Si data es wide se puede transformar con make_long (long format of data is required to get party labels with get_names. If data is in wide format you can transform it with make_long).

Examples

 tucuman_dip_gral_2017
#> # A tibble: 6 × 8
#> # Groups:   codprov [1]
#>   category round  year codprov name_prov electores listas   votos
#>   <chr>    <chr> <dbl> <chr>   <chr>         <dbl> <chr>    <dbl>
#> 1 dip      gral   2017 23      TUCUMAN     1217274 0180    154930
#> 2 dip      gral   2017 23      TUCUMAN     1217274 0503     46609
#> 3 dip      gral   2017 23      TUCUMAN     1217274 0521    319221
#> 4 dip      gral   2017 23      TUCUMAN     1217274 0548    459257
#> 5 dip      gral   2017 23      TUCUMAN     1217274 blancos   5920
#> 6 dip      gral   2017 23      TUCUMAN     1217274 nulos    12947

 tucuman_dip_gral_2017 %>%
     get_names()
#> # A tibble: 6 × 9
#> # Groups:   codprov [1]
#>   category round  year codprov name_prov electores listas   votos nombre_lista  
#>   <chr>    <chr> <dbl> <chr>   <chr>         <dbl> <chr>    <dbl> <chr>         
#> 1 dip      gral   2017 23      TUCUMAN     1217274 0180    154930 FUERZA REPUBL…
#> 2 dip      gral   2017 23      TUCUMAN     1217274 0503     46609 FRENTE DE IZQ…
#> 3 dip      gral   2017 23      TUCUMAN     1217274 0521    319221 CAMBIEMOS PAR…
#> 4 dip      gral   2017 23      TUCUMAN     1217274 0548    459257 FRENTE JUSTIC…
#> 5 dip      gral   2017 23      TUCUMAN     1217274 blancos   5920 blancos       
#> 6 dip      gral   2017 23      TUCUMAN     1217274 nulos    12947 nulos