Diccionario de elecciones disponibles (Elections collection)

Función que devuelve un tibble con los parámetros necesarios para llamar elecciones con get_election_data o get_election_results (Function that returns a tibble with the necessary parameters to call elections with get_election_data or get_election_results).

show_available_elections(source = NULL, viewer = FALSE)

Arguments

source

Fuente de consulta. Las opciones son 'data' para get_election_data y 'results' para get_election_results

viewer

Por defecto es FALSE. Cuando TRUE devuelve una tabla en el Viewer de RStudio (The default is FALSE. When TRUE it returns a table in RStudio Viewer).

Value

El objeto de salida es un data set con los id de elecciones necesarios como parámetros para usar con get_election_data. Cuando el parámetro es viewer = FALSE, devuelve un tibble con class "tbl_df","tbl","data.frame", y cuando es viewer = TRUE devuelve un objeto con class "datatables","htmlwidget"

(The output is a data set with elections id needed as parameters in get_election_data. When parameter is set to viewer = FALSE it returns a tibble and when it is viewer = TRUE it returns an object of class "datatables","htmlwidget").

Examples

 show_available_elections(source= 'data', viewer = FALSE)
#> # A tibble: 426 × 5
#>    district category round  year  NOMBRE   
#>    <chr>    <chr>    <chr>  <chr> <chr>    
#>  1 arg      presi    balota 2015  ARGENTINA
#>  2 arg      presi    gral   2003  ARGENTINA
#>  3 arg      presi    gral   2007  ARGENTINA
#>  4 arg      presi    gral   2011  ARGENTINA
#>  5 arg      presi    gral   2015  ARGENTINA
#>  6 arg      presi    gral   2019  ARGENTINA
#>  7 arg      presi    paso   2011  ARGENTINA
#>  8 arg      presi    paso   2015  ARGENTINA
#>  9 arg      presi    paso   2019  ARGENTINA
#> 10 caba     dip      gral   2005  CABA     
#> # ℹ 416 more rows