Descarga resultados de múltiples elecciones (Download multiple election electoral data)

Esta función permite descargar resultados de una multiplicidad de elecciones al mismo tiempo.

(Function that downloads multiple national electoral data with one call)

get_multiple_elections(
  data,
  source = NULL,
  unnest = FALSE,
  level = "provincia"
)

Arguments

data
data.frame con tantas filas como elecciones se quiere descargar y cuatro columnas con las siguientes variables: district, category, round, year

(data.frame with as many rows as elections you want to download and four columns with the following variables:district, category, round, year).

source
Fuente de los datos. Las opciones son 'data' para datosget_election_data y 'results' para get_election_results

unnest
un boleano que devuelve los datos anidados cuando TRUE agrupando cada elección o un data.frame cuando es FALSE que incluye una variable de id de la elección

(a boolean that returns nested data when TRUE, grouping each election or a data.frame when FALSE that includes an election id variable).

level
parámetro para definir el nivel de agregación de los datos que se quieren descargar ('provincia', 'departamento', 'circuito'). Por defecto es provincia

(parameter to define the level of aggregation of the data to be downloaded ('province', 'department', 'circuit'). Default is province).

Value

devuelve un tibble con class "grouped_df", "tbl_df","tbl", "data.frame" con los resultados de las elección seleccionadas, con tantas filas como elecciones se consultaron y dos columnas: id de la elección construido como concatenación de los parámetros year_category_round_year; election contiene un listado de tibbles con los resultados agregados a nivel provincial para cada elección

(returns a tibble of class "grouped_df", "tbl_df", "tbl", "data.frame" with as many rows as elections requested and two columns: id of the election build as a concatenation of the parameters year_category_round_year; election contains a list of tibbles with electoral results aggregated at the provincial level for each each row).

See also

Examples

# \donttest{
 electorAr::show_available_elections(source = 'data') %>%
 dplyr::filter(district == "caba",
              category == "dip",
              round == "paso") -> caba_paso_diputados

caba_paso_diputados

#> # A tibble: 5 x 5
#>   district category round year  NOMBRE
#>   <chr>    <chr>    <chr> <chr> <chr> 
#> 1 caba     dip      paso  2011  CABA  
#> 2 caba     dip      paso  2013  CABA  
#> 3 caba     dip      paso  2015  CABA  
#> 4 caba     dip      paso  2017  CABA  
#> 5 caba     dip      paso  2019  CABA  

 get_multiple_elections(data = caba_paso_diputados, source = 'data')

#> # A tibble: 5 x 2
#> # Groups:   id [5]
#>   id                 election                 
#>   <glue>             <list>                   
#> 1 caba_dip_paso_2011 <grouped_df[,8] [14 x 8]>
#> 2 caba_dip_paso_2013 <grouped_df[,8] [15 x 8]>
#> 3 caba_dip_paso_2015 <grouped_df[,8] [15 x 8]>
#> 4 caba_dip_paso_2017 <grouped_df[,8] [18 x 8]>
#> 5 caba_dip_paso_2019 <grouped_df[,8] [14 x 8]>

 # }